+++ mobile/android/android-components/components/feature/contextmenu/src/main/java/mozilla/components/feature/contextmenu/ContextMenuFragment.kt
@@ -7,7 +7,9 @@ package mozilla.components.feature.contextmenu
 import android.annotation.SuppressLint
 import android.app.Dialog
 import android.content.DialogInterface
+import android.os.Build
 import android.os.Bundle
+import android.text.Html
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -15,9 +17,11 @@ import android.widget.TextView
 import androidx.annotation.VisibleForTesting
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.widget.AppCompatTextView
+import androidx.core.text.HtmlCompat
 import androidx.fragment.app.DialogFragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.google.android.material.textview.MaterialTextView
 import mozilla.components.browser.state.state.SessionState
 
 private const val EXPANDED_TITLE_MAX_LINES = 15
@@ -25,6 +29,7 @@ private const val KEY_TITLE = "title"
 private const val KEY_SESSION_ID = "session_id"
 private const val KEY_IDS = "ids"
 private const val KEY_LABELS = "labels"
+private const val KEY_ADDITIONAL_NOTE = "additional_note"
 
 /**
  * [DialogFragment] implementation to display the actual context menu dialog.
@@ -46,6 +51,10 @@ class ContextMenuFragment : DialogFragment() {
 
     @VisibleForTesting internal val title: String by lazy {
         requireArguments().getString(KEY_TITLE)!!
+    }
+
+    @VisibleForTesting internal val additionalNote: String? by lazy {
+        requireArguments().getString(KEY_ADDITIONAL_NOTE)
     }
 
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
@@ -83,7 +92,20 @@ class ContextMenuFragment : DialogFragment() {
             adapter = ContextMenuAdapter(this@ContextMenuFragment, inflater)
         }
 
+        additionalNote?.let { value ->
+            val additionalNoteView = view.findViewById<MaterialTextView>(R.id.additional_note)
+            additionalNoteView.visibility = View.VISIBLE
+            additionalNoteView.text = getSpannedValueOfString(value)
+        }
+
         return view
+    }
+
+    private fun getSpannedValueOfString(value: String) = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+        Html.fromHtml(value, HtmlCompat.FROM_HTML_MODE_LEGACY)
+    } else {
+        @Suppress("Deprecation")
+        Html.fromHtml(value)
     }
 
     internal fun onItemSelected(position: Int) {
@@ -104,13 +126,15 @@ class ContextMenuFragment : DialogFragment() {
             tab: SessionState,
             title: String,
             ids: List<String>,
-            labels: List<String>
+            labels: List<String>,
+            additionalNote: String?
         ): ContextMenuFragment {
             val arguments = Bundle()
             arguments.putString(KEY_TITLE, title)
             arguments.putStringArrayList(KEY_IDS, ArrayList(ids))
             arguments.putStringArrayList(KEY_LABELS, ArrayList(labels))
             arguments.putString(KEY_SESSION_ID, tab.id)
+            arguments.putString(KEY_ADDITIONAL_NOTE, additionalNote)
 
             val fragment = ContextMenuFragment()
             fragment.arguments = arguments