+++ mobile/android/android-components/docs/changelog.md
@@ -8,6 +8,9 @@ permalink: /changelog/
   * 🌟️ Adds icons: mozac_ic_add_to_home_screen, mozac_ic_help, mozac_ic_shield, mozac_ic_shield_disabled
   * 🌟️ Update icons: mozac_ic_home, mozac_ic_settings, mozac_ic_clear
 
+* **feature-contextmenu**:
+  * 🌟️ Adds `additionalNote` which it will be attached to the bottom of context menu but for a specific `HitResult`
+
 # 93.0.0-SNAPSHOT (In Development)
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v92.0.0...main)
 * [Milestone](https://github.com/mozilla-mobile/android-components/milestone/140?closed=1)